+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - notification notifier url
+- locationModel identity optional
 
 ## 2.5.1 - 2023-07-31